Separate name steps for shared resources and app resources (#668)
