+++ src/commands/deployWorkspaceProject/internal/DeployWorkspaceProjectSaveSettingsStep.ts
@@ -12,7 +12,6 @@ import { createActivityChildContext } from "../../../utils/activity/activityUtil
 import { localize } from "../../../utils/localize";
 import { type DeploymentConfigurationSettings } from "../settings/DeployWorkspaceProjectSettingsV2";
 import { dwpSettingUtilsV2 } from "../settings/dwpSettingUtilsV2";
-import { containerAppSuffix } from "./DefaultResourcesNameStep";
 import { type DeployWorkspaceProjectInternalContext } from "./DeployWorkspaceProjectInternalContext";
 
 const saveSettingsLabel: string = localize('saveSettingsLabel', 'Save deployment settings to workspace "{0}"', relativeSettingsFilePath);
@@ -29,7 +28,7 @@ export class DeployWorkspaceProjectSaveSettingsStep extends ExecuteActivityOutpu
 
         const configurationLabel: string | undefined = context.configurationIdx !== undefined ? deploymentConfigurations?.[context.configurationIdx].label : undefined;
         const deploymentConfiguration: DeploymentConfigurationSettings = {
-            label: configurationLabel || removeCaSuffixIfExists(nonNullValueAndProp(context.containerApp, 'name')),
+            label: configurationLabel || nonNullValueAndProp(context.containerApp, 'name'),
             type: 'AcrDockerBuildRequest',
             dockerfilePath: path.relative(rootFolder.uri.fsPath, nonNullProp(context, 'dockerfilePath')),
             srcPath: path.relative(rootFolder.uri.fsPath, context.srcPath || rootFolder.uri.fsPath) || ".",
@@ -77,8 +76,4 @@ export class DeployWorkspaceProjectSaveSettingsStep extends ExecuteActivityOutpu
             message: localize('savedSettingsFail', 'Failed to save deployment settings to workspace "{0}".', relativeSettingsFilePath)
         };
     }
-}
-
-function removeCaSuffixIfExists(caName: string): string {
-    return caName.endsWith(containerAppSuffix) ? caName.slice(0, -containerAppSuffix.length) : caName;
 }
